Allow instantiation of Response base class
I currently have some legacy response models with type = "Response".
This is kinda invalid because it's defined as an abstract class.
For now, I want to be able to see these in Rails Admin without errors,
but in future if there are no invalid records in the DB, it could be
reverted.